fix : 지도에서 선택된 마커 오작동 버그 수정

+++ data/src/main/java/data/sahonmu/burger87/dto/announcement/AnnouncementDto.kt
@@ -12,6 +12,6 @@ data class AnnouncementDto(
     @SerialName("title") val title: String,
     @SerialName("contents") val contents: String,
     @SerialName("image") val image: String?,
-    @SerialName("store_id") val storeId: Int?,
+    @SerialName("store_id") val storeId: Long?,
     @SerialName("header") val isHeader: Boolean,
 )
